Added soft deletion option

+++ fn_exchange/fn_exchange/components/exchange_delete_emails.py
@@ -28,6 +28,7 @@ class FunctionComponent(ResilientComponent):
             # Get the function parameters:
             exchange_email = kwargs.get("exchange_email")  # text
             exchange_folder_path = kwargs.get("exchange_folder_path")  # text
+            exchange_hard_delete = kwargs.get("exchange_hard_delete") # boolean
             exchange_email_ids = kwargs.get("exchange_email_ids")  # text
             exchange_sender = kwargs.get("exchange_sender")  # text
             exchange_message_subject = kwargs.get("exchange_message_subject") # text
@@ -45,6 +46,7 @@ class FunctionComponent(ResilientComponent):
                 log.info('No folder path was specified, using value from config file')
             log.info("exchange_email: %s" % exchange_email)
             log.info("exchange_folder_path: %s" % exchange_folder_path)
+            log.info("exchange_hard_delete: %s" % exchange_hard_delete)
             log.info("exchange_email_ids: %s" % exchange_email_ids)
             log.info("exchange_sender: %s" % exchange_sender)
             log.info("exchange_message_subject: %s" % exchange_message_subject)
@@ -73,7 +75,11 @@ class FunctionComponent(ResilientComponent):
 
             # Delete Emails
             yield StatusMessage("Deleting emails")
-            emails.delete()
+            if exchange_hard_delete:
+                emails.delete()
+            else:
+                for item in emails:
+                    item.move_to_trash()
             yield StatusMessage("Done deleting emails, %d emails deleted" % num_deleted)
 
             # Produce a FunctionResult with the results
